Added Error Recorder to keep track failed processes
To keep track of failed processes and to handle retries. Retries on a
scheduler that we control rather than sidekiq. It should be made clear
that other services or classes would have to follow the same pattern
calls in order for the function to be invoked by the ErrorRecorder to
work. That is the reason why established styleguides and standardization
is warranted.